ci(cognizes): 修复 Backend Unit Tests 和 UI Quality CI 失败;
- setup-yarn-ui: CI 中临时移除根 packageManager 字段，解决 Node 22 Corepack
  检测到 pnpm 后拒绝执行 yarn 的问题，使 UI Quality 5 个 job 恢复正常
- 将 tests/unittests/engine/pulse/test_state_manager.py 移至
  tests/integration/，消除 unit test job 中 13 个 PostgreSQL 连接错误
- 将 TestPostgresSessionServiceIntegration 从
  tests/unittests/engine/mind/test_session_service.py 拆分至
  tests/integration/engine/mind/test_session_service.py，消除 4 个错误
